# metrics: make gauge_float64 and counter_float64 lock free