--- a/metrics/counter_float_64_test.go
+++ b/metrics/counter_float_64_test.go
@@ -1,6 +1,7 @@
 package metrics
 
 import (
+	"sync"
 	"testing"
 )
 
@@ -15,11 +16,20 @@ func BenchmarkCounterFloat64(b *testing.B) {
 func BenchmarkCounterFloat64Parallel(b *testing.B) {
 	c := NewCounterFloat64()
 	b.ResetTimer()
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			c.Inc(1.0)
-		}
-	})
+	var wg sync.WaitGroup
+	for i := 0; i < 10; i++ {
+		wg.Add(1)
+		go func() {
+			for i := 0; i < b.N; i++ {
+				c.Inc(1.0)
+			}
+			wg.Done()
+		}()
+	}
+	wg.Wait()
+	if have, want := c.Count(), 10.0*float64(b.N); have != want {
+		b.Fatalf("have %f want %f", have, want)
+	}
 }
 
 func TestCounterFloat64Clear(t *testing.T) {