--- a/metrics/gauge_float64.go
+++ b/metrics/gauge_float64.go
@@ -79,7 +79,7 @@ func (NilGaugeFloat64) Update(v float64) {}
 func (NilGaugeFloat64) Value() float64 { return 0.0 }
 
 // StandardGaugeFloat64 is the standard implementation of a GaugeFloat64 and uses
-// sync.Mutex to manage a single float64 value.
+// atomic to manage a single float64 value.
 type StandardGaugeFloat64 struct {
 	floatBits uint64
 }