--- a/metrics/gauge_float64_test.go
+++ b/metrics/gauge_float64_test.go
@@ -1,6 +1,7 @@
 package metrics
 
 import (
+	"sync"
 	"testing"
 )
 
@@ -14,14 +15,20 @@ func BenchmarkGaugeFloat64(b *testing.B) {
 
 func BenchmarkGaugeFloat64Parallel(b *testing.B) {
 	c := NewGaugeFloat64()
-	b.ResetTimer()
-	b.RunParallel(func(pb *testing.PB) {
-		count := 1.0
-		for pb.Next() {
-			c.Update(count)
-			count += 1.0
-		}
-	})
+	var wg sync.WaitGroup
+	for i := 0; i < 10; i++ {
+		wg.Add(1)
+		go func() {
+			for i := 0; i < b.N; i++ {
+				c.Update(float64(i))
+			}
+			wg.Done()
+		}()
+	}
+	wg.Wait()
+	if have, want := c.Value(), float64(b.N-1); have != want {
+		b.Fatalf("have %f want %f", have, want)
+	}
 }
 
 func TestGaugeFloat64(t *testing.T) {