This PR changes the gauge and counter of type f64 to use atomic instead of mutex.

BenchmarkCounterFloat64Pararel-32       186.0 ns/op
BenchmarkCounterFloat64Pararel-32       169.3 ns/op

BenchmarkGaugeFloat64Pararel-32         180.7 ns/op
BenchmarkGaugeFloat64Pararel-32         170.2 ns/op