@holiman 2023-04-02
```
[user@work metrics]$ benchstat float.1 float.2
name              old time/op  new time/op  delta
CounterFloat64-8  36.1ns ± 7%  21.2ns ±11%  -41.45%  (p=0.008 n=5+5)
GaugeFloat64-8    32.3ns ±27%  14.3ns ±23%  -55.63%  (p=0.008 n=5+5)
```
:+1: