@holiman 2023-04-03
I pushed some changes now, removing some very small utility methods which I just inlined instead. The way you had written the benchmark, using runParallel doesn't seem to work quite as one might expect. When I tested that before vs after, it showed a degradation. 
I made a more straight-forward benchmark instead, which just mimics ten writers trying to update the counter. That bench shows improvement as expected

```
name                      old time/op  new time/op  delta
CounterFloat64Parallel-8  1.45µs ±10%  0.85µs ± 6%  -41.65%  (p=0.008 n=5+5)

```